Feat: moved table to a different file
